# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BSV session settlement MOCK",
    "description": "Simulated payments only. No blockchain transactions.",
    "version": "0.1.0"
  },
  "paths": {
    "/v1/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_v1_health_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/wallet-bindings/{binding_id}": {
      "get": {
        "summary": "Binding",
        "operationId": "binding_v1_wallet_bindings__binding_id__get",
        "parameters": [
          {
            "name": "binding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Binding Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/settlements/{sid}": {
      "put": {
        "summary": "Prepare",
        "operationId": "prepare_v1_settlements__sid__put",
        "parameters": [
          {
            "name": "sid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Sid"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Prepare"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Status",
        "operationId": "status_v1_settlements__sid__get",
        "parameters": [
          {
            "name": "sid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Sid"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/settlements/{sid}/request-payment": {
      "post": {
        "summary": "Request Payment",
        "operationId": "request_payment_v1_settlements__sid__request_payment_post",
        "parameters": [
          {
            "name": "sid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Sid"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPayment"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/settlements/{sid}/quotes": {
      "post": {
        "summary": "Replace Quote",
        "operationId": "replace_quote_v1_settlements__sid__quotes_post",
        "parameters": [
          {
            "name": "sid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Sid"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Requote"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/mock/settlements/{sid}/decision": {
      "post": {
        "summary": "Decide",
        "operationId": "decide_v1_mock_settlements__sid__decision_post",
        "parameters": [
          {
            "name": "sid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Sid"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Decision"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/mock/settlements/{sid}/confirm": {
      "post": {
        "summary": "Confirm",
        "operationId": "confirm_v1_mock_settlements__sid__confirm_post",
        "parameters": [
          {
            "name": "sid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Sid"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Decision": {
        "properties": {
          "approval_request_id": {
            "type": "string",
            "title": "Approval Request Id"
          },
          "quote_id": {
            "type": "string",
            "title": "Quote Id"
          },
          "decision": {
            "type": "string",
            "enum": [
              "approve",
              "decline"
            ],
            "title": "Decision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "approval_request_id",
          "quote_id",
          "decision"
        ],
        "title": "Decision"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Prepare": {
        "properties": {
          "session_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Session Id"
          },
          "ledger_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Ledger Revision",
            "default": 1
          },
          "ledger_sha256": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$",
            "title": "Ledger Sha256"
          },
          "driver_binding_id": {
            "type": "string",
            "const": "driver-demo-01",
            "title": "Driver Binding Id"
          },
          "operator_binding_id": {
            "type": "string",
            "const": "operator-demo-01",
            "title": "Operator Binding Id"
          },
          "currency": {
            "type": "string",
            "const": "AUD",
            "title": "Currency"
          },
          "net_amount_minor": {
            "type": "integer",
            "maximum": 100000.0,
            "minimum": -100000.0,
            "title": "Net Amount Minor"
          },
          "import_wh": {
            "type": "integer",
            "maximum": 10000000.0,
            "minimum": 0.0,
            "title": "Import Wh"
          },
          "export_wh": {
            "type": "integer",
            "maximum": 10000000.0,
            "minimum": 0.0,
            "title": "Export Wh"
          },
          "pricing_summary": {
            "$ref": "#/components/schemas/Summary"
          },
          "ended_at": {
            "type": "string",
            "format": "date-time",
            "title": "Ended At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id",
          "ledger_sha256",
          "driver_binding_id",
          "operator_binding_id",
          "currency",
          "net_amount_minor",
          "import_wh",
          "export_wh",
          "pricing_summary",
          "ended_at"
        ],
        "title": "Prepare"
      },
      "RequestPayment": {
        "properties": {
          "quote_id": {
            "type": "string",
            "title": "Quote Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "quote_id"
        ],
        "title": "RequestPayment"
      },
      "Requote": {
        "properties": {
          "replaces_quote_id": {
            "type": "string",
            "title": "Replaces Quote Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "replaces_quote_id"
        ],
        "title": "Requote"
      },
      "Summary": {
        "properties": {
          "import_amount_aud": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Import Amount Aud"
          },
          "export_credit_aud": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Export Credit Aud"
          },
          "rounding_adjustment_aud": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Rounding Adjustment Aud"
          },
          "prices_final": {
            "type": "boolean",
            "const": true,
            "title": "Prices Final"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "import_amount_aud",
          "export_credit_aud",
          "rounding_adjustment_aud",
          "prices_final"
        ],
        "title": "Summary"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}